@@ -1,9 +1,14 @@
 {
     "name": "doctrine/collections",
     "type": "library",
-    "description": "Collections Abstraction library",
-    "keywords": ["collections", "array", "iterator"],
-    "homepage": "http://www.doctrine-project.org",
+    "description": "PHP Doctrine Collections library that adds additional functionality on top of PHP arrays.",
+    "keywords": [
+        "php",
+        "collections",
+        "array",
+        "iterators"
+    ],
+    "homepage": "https://www.doctrine-project.org/projects/collections.html",
     "license": "MIT",
     "authors": [
         {"name": "Guilherme Blanco", "email": "guilhermeblanco@gmail.com"},
@@ -13,17 +18,25 @@
         {"name": "Johannes Schmitt", "email": "schmittjoh@gmail.com"}
     ],
     "require": {
-        "php": ">=5.3.2"
+        "php": "^7.1.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "phpunit/phpunit": "^7.0",
+        "doctrine/coding-standard": "^6.0",
+        "phpstan/phpstan-shim": "^0.9.2",
+        "vimeo/psalm": "^3.2.2"
     },
     "autoload": {
-        "psr-0": { "Doctrine\\Common\\Collections\\": "lib/" }
+        "psr-4": { "Doctrine\\Common\\Collections\\": "lib/Doctrine/Common/Collections" }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Doctrine\\Tests\\": "tests/Doctrine/Tests"
+        }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.2.x-dev"
+            "dev-master": "1.6.x-dev"
         }
     }
 }
