@@ -3,7 +3,7 @@
     "description":       "A small, lightweight utility to instantiate objects in PHP without invoking their constructors",
     "type":              "library",
     "license":           "MIT",
-    "homepage":          "https://github.com/doctrine/instantiator",
+    "homepage":          "https://www.doctrine-project.org/projects/instantiator.html",
     "keywords":          [
         "instantiate",
         "constructor"
@@ -16,14 +16,16 @@
         }
     ],
     "require": {
-        "php": ">=5.3,<8.0-DEV"
+        "php": "^7.1"
     },
     "require-dev": {
         "ext-phar":                  "*",
         "ext-pdo":                   "*",
-        "phpunit/phpunit":           "~4.0",
-        "squizlabs/php_codesniffer": "~2.0",
-        "athletic/athletic":         "~0.1.8"
+        "doctrine/coding-standard":  "^6.0",
+        "phpbench/phpbench":         "^0.13",
+        "phpstan/phpstan-phpunit":   "^0.11",
+        "phpstan/phpstan-shim":      "^0.11",
+        "phpunit/phpunit":           "^7.0"
     },
     "autoload": {
         "psr-4": {
@@ -39,7 +41,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0.x-dev"
+            "dev-master": "1.2.x-dev"
         }
     }
 }
