@@ -1,26 +1,34 @@
 {
     "name":    "phpdocumentor/reflection-docblock",
+    "description": "With this component, a library can provide support for annotations via DocBlocks or otherwise retrieve information that is embedded in a DocBlock.",
     "type":    "library",
     "license": "MIT",
     "authors": [
-        {"name": "Mike van Riel", "email": "mike.vanriel@naenius.com"}
+        {
+            "name": "Mike van Riel",
+            "email": "me@mikevanriel.com"
+        }
     ],
     "require": {
-        "php": ">=5.3.3"
+        "php": "^7.0",
+        "phpdocumentor/reflection-common": "^1.0.0 || ^2.0.0",
+        "phpdocumentor/type-resolver": "~0.4 || ^1.0.0",
+        "webmozart/assert": "^1.0"
     },
     "autoload": {
-        "psr-0": {"phpDocumentor": ["src/"]}
+        "psr-4": {"phpDocumentor\\Reflection\\": ["src/"]}
+    },
+    "autoload-dev": {
+        "psr-4": {"phpDocumentor\\Reflection\\": ["tests/unit"]}
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0"
-    },
-    "suggest": {
-        "dflydev/markdown": "~1.0",
-        "erusev/parsedown": "~1.0"
+        "mockery/mockery": "^1.0",
+        "phpunit/phpunit": "^6.4",
+        "doctrine/instantiator": "^1.0.5"
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.0.x-dev"
+            "dev-master": "4.x-dev"
         }
     }
 }
