@@ -7,7 +7,7 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 /**
  * Listens to the dynamic route events.
  */
-class socialMediaSubscriber implements EventSubscriberInterface {
+class SocialMediaSubscriber implements EventSubscriberInterface {
 
   /**
    * Public static method used by Drupals event.
@@ -16,18 +16,16 @@ class socialMediaSubscriber implements EventSubscriberInterface {
    *   Array of events we subscribe to.
    */
   public static function getSubscribedEvents() {
-    $events = array();
-    $events['social_media.add_more_social_media'][] = array('addMoreSocialMedia', 39);
-    $events['social_media.pre_execute'][] = array('pre_executeSocialMedia', 39);
-    $events['social_media.pre_render'][] = array('pre_rednerSocialMedia', 39);
+    $events = [];
+    $events['social_media.add_more_social_media'][] = ['addMoreSocialMedia', 39];
+    $events['social_media.pre_execute'][] = ['executeSocialMedia', 39];
+    $events['social_media.pre_render'][] = ['renderSocialMedia', 39];
 
     return $events;
   }
 
-
   /**
    * Subscribes to  social_media.add_more_social_media event.
-   *
    */
   public function addMoreSocialMedia($event) {
     $element = $event->getElement();
@@ -37,20 +35,17 @@ class socialMediaSubscriber implements EventSubscriberInterface {
 
   /**
    * Subscribes to social_media.pre_execute event.
-   *
    */
-  public function pre_executeSocialMedia($event) {
+  public function executeSocialMedia($event) {
     $element = $event->getElement();
     $element['facebook_share']['weight'] = 8;
     $event->setElement($element);
   }
 
-
   /**
    * Subscribes to social_media.pre_render event.
-   *
    */
-  public function pre_rednerSocialMedia($event) {
+  public function renderSocialMedia($event) {
     $element = $event->getElement();
     $element['facebook_msg']['text'] = 'New facebook msg text';
     $event->setElement($element);
